# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RSS AI Digest API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /metrics:
    get:
      summary: Export minimal Prometheus metrics
      responses:
        '200':
          description: Prometheus text format
          content:
            text/plain:
              schema:
                type: string
  /api/v1/articles:
    get:
      summary: List articles
      responses:
        '200':
          description: Article list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
        '503':
          description: Article reader is not configured
  /api/v1/digests/latest:
    get:
      summary: Get latest digest
      responses:
        '200':
          description: Latest digest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Digest'
        '503':
          description: Digest reader is not configured
  /api/v1/profiles/{profileType}/active:
    get:
      summary: Get active profile by type
      parameters:
        - in: path
          name: profileType
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Active profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '503':
          description: Profile reader is not configured
  /api/v1/jobs/daily-digest:
    post:
      summary: Trigger daily digest job
      security:
        - apiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                trigger_at:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Job skipped because the same digest date is already queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTriggerResponse'
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTriggerResponse'
        '400':
          description: Invalid request body or trigger_at
        '401':
          description: Invalid API key
        '500':
          description: Job enqueue failed
        '503':
          description: Job trigger is not configured
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Article:
      type: object
      properties:
        id:
          type: string
        miniflux_entry_id:
          type: integer
          format: int64
        feed_id:
          type: integer
          format: int64
        feed_title:
          type: string
        title:
          type: string
        author:
          type: string
        url:
          type: string
          format: uri
        title_translated:
          type: string
        summary_translated:
          type: string
        content_translated:
          type: string
        core_summary:
          type: string
        key_points:
          type: array
          items:
            type: string
        topic_category:
          type: string
        importance_score:
          type: number
          format: float
    Digest:
      type: object
      properties:
        digest_date:
          type: string
          format: date
        title:
          type: string
        subtitle:
          type: string
        content_markdown:
          type: string
        content_html:
          type: string
        remote_url:
          type: string
        publish_state:
          type: string
        publish_error:
          type: string
    Profile:
      type: object
      properties:
        profile_type:
          type: string
        name:
          type: string
        version:
          type: integer
        is_active:
          type: boolean
        payload:
          type: object
          additionalProperties: true
    JobTriggerResponse:
      type: object
      properties:
        digest_date:
          type: string
          format: date
        status:
          type: string
          enum: [accepted, skipped]
        trigger_at:
          type: string
          format: date-time